+++ examples/next-minimal-starter/src/utils/trpc.ts
@@ -4,10 +4,10 @@ import type { AppRouter } from '../pages/api/trpc/[trpc]';
 export const trpc = createReactQueryHooks<AppRouter>();
 // => { useQuery: ..., useMutation: ...}
 */
-import { setupTRPC } from '@trpc/next';
+import { createTRPCNext } from '@trpc/next';
 import { AppRouter } from '../pages/api/trpc/[trpc]';
 
-export const trpc = setupTRPC<AppRouter>({
+export const trpc = createTRPCNext<AppRouter>({
   config() {
     return {
       url: '/api/trpc',
