+++ packages/react/src/shared/index.ts
@@ -1,3 +1,4 @@
 export * from './proxy/decorationProxy';
 export * from './proxy/utilsProxy';
 export type { DecoratedProcedureRecord } from '../createTRPCReact';
+export * from './hooks/createHooksInternal';
